Much simplified testing * easier to define tests using Util.TestCase * simple example of usage in test/testCollection.py
